requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi
uvicorn[standard]
pydantic[dotenv]
python-multipart
loguru
rich
tqdm
pandas
numpy
scipy
scikit-learn
torch
torchvision
torchaudio
transformers
accelerate
sentence-transformers
faiss-cpu
langchain
chromadb
pymupdf
pdfminer.six
camelot-py
tabula-py
pytesseract
pillow
opencv-python
spacy
pymoo
botorch
gpytorch
plotly
streamlit
mlflow
dvc
pyyaml
python-dotenv
joblib

